Add files via upload
This is a system written in the language c. it is a bus reservation system that can fulfill the following. to start with it can allow the signing up of employees and logging in of employees. One can also add a bus schedule and view a bus schedule. There is also a function to book a seat and cancel a booking.